+++ lib/src/importer/js_to_dart/async.dart
@@ -2,49 +2,44 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
+import 'dart:async';
+
 import 'package:node_interop/js.dart';
+import 'package:node_interop/util.dart';
 
-import '../../importer.dart';
-import '../../node/importer.dart';
-import '../../node/url.dart';
-import '../../node/utils.dart';
+import '../../js/importer.dart';
+import '../../js/url.dart';
+import '../../js/utils.dart';
 import '../../util/nullable.dart';
+import '../async.dart';
+import '../result.dart';
 
-/// A wrapper for a synchronous JS API importer that exposes it as a Dart
-/// [Importer].
-final class NodeToDartImporter extends Importer {
+/// A wrapper for a potentially-asynchronous JS API importer that exposes it as
+/// a Dart [AsyncImporter].
+final class JSToDartAsyncImporter extends AsyncImporter {
   /// The wrapped canonicalize function.
   final Object? Function(String, CanonicalizeOptions) _canonicalize;
 
   /// The wrapped load function.
   final Object? Function(JSUrl) _load;
 
-  NodeToDartImporter(this._canonicalize, this._load);
+  JSToDartAsyncImporter(this._canonicalize, this._load);
 
-  Uri? canonicalize(Uri url) {
+  FutureOr<Uri?> canonicalize(Uri url) async {
     var result = _canonicalize(
         url.toString(), CanonicalizeOptions(fromImport: fromImport));
+    if (isPromise(result)) result = await promiseToFuture(result as Promise);
     if (result == null) return null;
+
     if (isJSUrl(result)) return jsToDartUrl(result as JSUrl);
 
-    if (isPromise(result)) {
-      jsThrow(JsError(
-          "The canonicalize() function can't return a Promise for synchronous "
-          "compile functions."));
-    } else {
-      jsThrow(JsError("The canonicalize() method must return a URL."));
-    }
+    jsThrow(JsError("The canonicalize() method must return a URL."));
   }
 
-  ImporterResult? load(Uri url) {
+  FutureOr<ImporterResult?> load(Uri url) async {
     var result = _load(dartToJSUrl(url));
+    if (isPromise(result)) result = await promiseToFuture(result as Promise);
     if (result == null) return null;
-
-    if (isPromise(result)) {
-      jsThrow(JsError(
-          "The load() function can't return a Promise for synchronous compile "
-          "functions."));
-    }
 
     result as NodeImporterResult;
     var contents = result.contents;