+++ test/dart_api/value/function_test.dart
@@ -29,6 +29,7 @@ void main() {
 
     test("isn't any other type", () {
       expect(value.assertBoolean, throwsSassScriptException);
+      expect(value.assertCalculation, throwsSassScriptException);
       expect(value.assertColor, throwsSassScriptException);
       expect(value.assertMap, throwsSassScriptException);
       expect(value.tryMap(), isNull);